pyproject.toml
[tool.poetry]
name = "dentistry"
version = "0.1.0"
description = "This is an API for a project to help dentists manage their clinics."
authors = ["M <m@mathsistor.com>", "Sika <sikabarca@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.0.6"
psycopg2-binary = "^2.8.5"
dj-database-url = "^0.5.0"
django-environ = "^0.4.5"
django-extensions = "^2.2.9"
djangorestframework = "^3.11.0"
django-cors-headers = "^3.2.1"
gunicorn = "^20.0.4"
whitenoise = "^5.0.1"
django-storages = "^1.9.1"
boto3 = "^1.13.11"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^4.0.0"}

[tool.poetry.dev-dependencies]
pytest-django = "^3.9.0"
notebook = "^6.0.3"
factory_boy = "^2.12.0"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
